added order and address flow

+++ src/main/java/com/niraj/ecommerce/model/OrderItem.java
@@ -28,5 +28,5 @@ public class OrderItem {
     private Long quantity;
 
     @Column(nullable = false)
-    private Long priceAtPurchase;
+    private Double priceAtPurchase;
 }
